# apps/backend/main/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/workspaces/{workspace_id}/pages/{page_id}/tasks/{task_id}": {
            "get": {
                "description": "Get a task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postgres.Task"
                        }
                    },
                    "401": {
                        "description": "Unauthorized if session is missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Update a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task object that needs to be updated",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tasksRouter.updateTaskBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postgres.Task"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized if session is missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Delete a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/postgres.Task"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    }
                }
            }
        },
...
        "/workspaces/{workspace_id}/pages/{page_id}/tasks/{task_id}/history": {
            "get": {
                "description": "Get task history",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task history",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page ID",
                        "name": "page_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Workspace ID",
                        "name": "workspace_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "task_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "If not provided or less than 1, all users will be returned",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Default is 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/routerUtils.ResponseWithPagination-mongo_EntityVersionDocument"
                        }
                    },
                    "401": {
                        "description": "Unauthorized if session is missing or invalid",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errorHandlers.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "errorHandlers.Error": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                },
                "errorCode": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "mongo.Change": {
            "type": "object",
            "properties": {
                "from": {},
                "to": {}
            }
        },
        "mongo.EntityVersionDocument": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/mongo.ShortUserInfo"
                },
                "changes": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/mongo.Change"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "mongo.ShortUserInfo": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "postgres.Page": {
            "type": "object",
            "properties": {
                "childPages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postgres.Page"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "owner": {
                    "$ref": "#/definitions/postgres.User"
                },
                "parentPage": {
                    "$ref": "#/definitions/postgres.Page"
                },
                "role": {
                    "$ref": "#/definitions/postgres.UserRole"
                },
                "tasks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/postgres.Task"
                    }
                },
                "text": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/postgres.PageType"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "postgres.PageType": {
            "type": "string",
            "enum": [
                "board",
                "text",
                "group"
            ],
            "x-enum-varnames": [
                "PageTypeBoard",
                "PageTypeText",
                "PageTypeGroup"
            ]
        },
        "postgres.Task": {
            "type": "object",
            "properties": {
                "assignee": {
                    "$ref": "#/definitions/postgres.User"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "dueDate": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "page": {
                    "$ref": "#/definitions/postgres.Page"
                },
                "reporter": {
                    "$ref": "#/definitions/postgres.User"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "postgres.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "emailVerified": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "lastLogin": {
                    "type": "string"
                },
                "lastPasswordReset": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "postgres.UserRole": {
            "type": "string",
            "enum": [
                "owner",
                "admin",
                "member",
                "commentator",
                "guest"
            ],
            "x-enum-varnames": [
                "UserRoleOwner",
                "UserRoleAdmin",
                "UserRoleMember",
                "UserRoleCommentator",
                "UserRoleGuest"
            ]
        },
...
        "routerUtils.Meta": {
            "type": "object",
            "properties": {
                "hasMore": {
                    "type": "boolean"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "routerUtils.ResponseWithPagination-mongo_EntityVersionDocument": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mongo.EntityVersionDocument"
                    }
                },
                "meta": {
                    "$ref": "#/definitions/routerUtils.Meta"
                }
            }
        },
...
        "tasksRouter.updateTaskBody": {
            "type": "object",
            "properties": {
                "assigneeId": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 255
                },
                "dueDate": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "not_done",
                        "in_progress",
                        "done"
                    ]
                },
                "title": {
                    "type": "string",
                    "maxLength": 63
                }
            }
        },
...
    }
}